scripts/requirements.txt
openai
python-dotenv
pymupdf
tiktoken
mammoth
